完善入口组与客户端配置恢复并清理仓库产物

+++ NetTest.App/ViewModels/MainWindowViewModel.Reporting.Support.cs
@@ -22,7 +22,7 @@ public sealed partial class MainWindowViewModel
         string Source,
         string? Name,
         string? BaseUrl,
-        int? Score,
+        double? Score,
         string Summary);
 
     private sealed record TrendSummarySnapshot(
